# backend/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        description: |-
          Error is a human-readable, user-safe message. For 5xx it is generic —
          internals are only ever written to the server log.
        example: file "assets.txt" must have a .csv extension
        type: string
      line:
        description: Line is the 1-indexed source line of a structurally malformed
          CSV, when known.
        example: 42
        type: integer
      trace_id:
        description: |-
          TraceID matches the X-Trace-Id response header and the server log lines
          for this request; quote it when reporting a 5xx.
        example: 7f6c1a52-3c1e-4a51-9d0b-5b1b2f7a9c11
        type: string
    required:
    - error
    type: object
...
  dto.ImportListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.ImportSummary'
        type: array
      limit:
        example: 25
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 3
        type: integer
    required:
    - items
    - limit
    - offset
    - total
    type: object
  dto.ImportPreviewResponse:
    properties:
      fingerprint:
        description: |-
          Fingerprint is a SHA-256 hex digest of the file's cells and of the
          validation outcome.
        example: 9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08
        type: string
      ignored_columns:
        description: IgnoredColumns lists header columns the importer does not recognise.
        example:
        - legacy_ref
        - notes
        items:
          type: string
        type: array
      importable_rows:
        example: 206
        type: integer
      rejected_rows:
        example: 16
        type: integer
      rejections:
        items:
          $ref: '#/definitions/dto.Rejection'
        type: array
      total_rows:
        example: 222
        type: integer
    required:
    - fingerprint
    - ignored_columns
    - importable_rows
    - rejected_rows
    - rejections
    - total_rows
    type: object
  dto.ImportResponse:
    properties:
      committed:
        example: false
        type: boolean
      ignored_columns:
        description: |-
          IgnoredColumns lists header columns the importer does not recognise;
          they are skipped (not an error) and not stored.
        example:
        - legacy_ref
        - notes
        items:
          type: string
        type: array
      import_id:
        example: 0b5f3c7e-8a52-4a55-9b0e-0c9f8a3f6d11
        format: uuid
        type: string
      imported_rows:
        example: 0
        type: integer
      rejected_rows:
        example: 8
        type: integer
      rejections:
        items:
          $ref: '#/definitions/dto.Rejection'
        type: array
      total_rows:
        example: 222
        type: integer
    required:
    - committed
    - ignored_columns
    - import_id
    - imported_rows
    - rejected_rows
    - rejections
    - total_rows
    type: object
  dto.ImportSummary:
    properties:
      client_ip:
        example: 203.0.113.7
        type: string
      committed:
        example: true
        type: boolean
      created_at:
        example: "2026-09-20T08:15:00Z"
        format: date-time
        type: string
      filename:
        example: grid_assets.csv
        type: string
      import_id:
        example: 0b5f3c7e-8a52-4a55-9b0e-0c9f8a3f6d11
        format: uuid
        type: string
      imported_rows:
        example: 206
        type: integer
      rejected_rows:
        example: 16
        type: integer
      total_rows:
        example: 222
        type: integer
    required:
    - client_ip
    - committed
    - created_at
    - filename
    - import_id
    - imported_rows
    - rejected_rows
    - total_rows
    type: object
  dto.LookupsResponse:
    properties:
      asset_types:
        example:
        - SUBSTATION
        - TRANSFORMER
        items:
          type: string
        type: array
      operational_statuses:
        example:
        - IN_SERVICE
        - MAINTENANCE
        items:
          type: string
        type: array
      parent_rules:
        items:
          $ref: '#/definitions/dto.ParentRule'
        type: array
      root_types:
        description: 'RootTypes are asset types with no parent rule: they must have
          no parent.'
        example:
        - SUBSTATION
        items:
          type: string
        type: array
    required:
    - asset_types
    - operational_statuses
    - parent_rules
    - root_types
    type: object
  dto.ParentRule:
    properties:
      child_type:
        example: TRANSFORMER
        type: string
      parent_type:
        example: SUBSTATION
        type: string
    required:
    - child_type
    - parent_type
    type: object
  dto.Rejection:
    properties:
      asset_id:
        example: TX-NR
        type: string
      csv_row:
        description: |-
          CSVRow is the row's original 1-indexed line in the uploaded file
          (the header is line 1), not its position after any reordering.
        example: 178
        type: integer
      reason:
        description: Reason holds every problem found on the row, joined with "; ".
        example: rating_kva cannot be negative (-500)
        type: string
    required:
    - asset_id
    - csv_row
    - reason
    type: object
...
info:
  contact: {}
  description: CSV asset import/validation and asset hierarchy explorer API.
  title: Smart Grid Asset Management API
  version: "1.0"
paths:
...
  /api/imports:
    get:
      description: 'One page of import activity, newest first: file name, client IP,
        row counts and whether data was committed. Imports only — deletions are not
        listed. Fetch `GET /api/imports/{importId}` for a run''s rejected rows.'
      parameters:
      - default: 25
        description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Rows to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ImportListResponse'
        "400":
          description: Bad limit or offset
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid x-api-key
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List past imports
      tags:
      - imports
    post:
      consumes:
      - multipart/form-data
      description: |-
        Uploads a CSV of grid assets, validates every row again, and stores the rows that pass in one transaction. Rejected rows, and any row beneath a rejected parent, are skipped and reported; rows whose `asset_id` already exists are rejected, never updated. `committed` is true when at least one row was stored.

        To let a user review the outcome first, call `POST /api/imports/preview` and send its `fingerprint` here as `expected_fingerprint`: if the file or the stored data changed since, nothing is written and the response is **412**. Without `expected_fingerprint` the rows that pass are stored directly.

        **Columns are matched by header name**, in any order (case, spaces, `-` and `_` are ignored: `Asset ID` = `asset_id`). Required: `asset_id`, `asset_type`, `asset_name`, `operational_status`. Optional (may be absent from the file or blank): `parent_asset_id`, `voltage_kv`, `rating_kva`, `manufacturer`, `model`, `serial_number`, `commissioned_date` (`YYYY-MM-DD`, `D/M/YYYY` or `D/M/YY`, day first). Unknown columns are ignored and listed in `ignored_columns`. Allowed asset types, statuses and parent/child pairings are read from the database — see `GET /api/lookups`.

        A file that is readable but has bad rows returns **200** with `rejections[]`; `csv_row` is the row's line in the uploaded file (header = line 1). A file that cannot be processed at all (not CSV, empty, required column missing) returns a 4xx `ErrorResponse` instead.
      parameters:
      - description: CSV file with a header row
        in: formData
        name: file
        required: true
        type: file
      - description: Fingerprint from POST /api/imports/preview for this file
        in: formData
        name: expected_fingerprint
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Import outcome — check `committed`, `imported_rows` and `rejections`
          schema:
            $ref: '#/definitions/dto.ImportResponse'
        "400":
          description: No file supplied, or the file is empty / has no data rows
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid x-api-key
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Another writer changed the stored data during the commit; nothing
            was stored, retry
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "412":
          description: expected_fingerprint no longer matches; nothing was stored,
            preview again
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "413":
          description: File larger than the configured upload limit
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "415":
          description: File does not have a .csv extension
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Malformed CSV (see `line`), or the header lacks a required
            column
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Unexpected error; quote `trace_id` when reporting
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Import assets from a CSV file
      tags:
      - imports
  /api/imports/{importId}:
    get:
      description: Re-fetches the result of an earlier import, including every rejected
        row. `ignored_columns` is only reported on the upload response and is empty
        here.
      parameters:
      - description: Import id (UUID) from POST /api/imports
        format: uuid
        in: path
        name: importId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ImportResponse'
        "400":
          description: importId is not a UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid x-api-key
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: No such import
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a past import
      tags:
      - imports
  /api/imports/preview:
    post:
      consumes:
      - multipart/form-data
      description: |-
        Validates the file exactly as `POST /api/imports` would and reports how many rows can be imported and which are rejected and why, **without storing anything** (no assets, no import record). Show this to the user, then confirm by sending the same file to `POST /api/imports` with `expected_fingerprint` set to the `fingerprint` returned here.

        Column rules, date formats (`YYYY-MM-DD`, `D/M/YYYY`, `D/M/YY`, day first) and structural errors are the same as for the import.
      parameters:
      - description: CSV file with a header row
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Validation outcome; nothing was stored
          schema:
            $ref: '#/definitions/dto.ImportPreviewResponse'
        "400":
          description: No file supplied, or the file is empty / has no data rows
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid x-api-key
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "413":
          description: File larger than the configured upload limit
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "415":
          description: File does not have a .csv extension
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Malformed CSV (see `line`), or the header lacks a required
            column
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Unexpected error; quote `trace_id` when reporting
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Check a CSV file without importing it
      tags:
      - imports
  /api/imports/template:
    get:
      description: A CSV with every importable column in its header and a few fictional
        example rows (a small valid hierarchy) showing how parents, dates and statuses
        are written. Delete the example rows before importing your own data.
      produces:
      - text/csv
      responses:
        "200":
          description: asset_import_template.csv
          schema:
            type: file
        "401":
          description: Missing or invalid x-api-key
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Download the import template
      tags:
      - imports
  /api/lookups:
    get:
      description: The asset types, operational statuses and parent/child type pairings
        that imports are validated against, read from the database (refreshed on startup
        and on every import). Use it to build filters and dropdowns instead of hardcoding
        values.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LookupsResponse'
        "401":
          description: Missing or invalid x-api-key
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List reference data
      tags:
      - lookups
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: API key (see API_KEY in .env). Required by every /api endpoint (sent
      as the x-api-key header). Only /healthz and these docs are open. This is a shared
      key, not user authentication (see ARCHITECTURE.md).
    in: header
    name: x-api-key
    type: apiKey
swagger: "2.0"